KW-192/refactor: AuthService & JwtTokenService 인터페이스 작성
